requirements.txt
python-dotenv
trafilatura
beautifulsoup4
langchain
requests
supabase
pandas
openai
pypdf
tiktoken
faiss-cpu
gradio
tqdm
pycryptodome
chromadb
supabase
pytest-playwright
unstructured[local-inference]
jupyter
streamlit
streamlit-extras
llama-index